requirements.txt
anyio==4.5.2
click==8.1.7
et-xmlfile==1.1.0
exceptiongroup==1.2.2
fastapi==0.95.2
fastapi-login==1.10.2
greenlet==3.1.1
h11==0.14.0
httptools==0.6.4
idna==3.10
Jinja2==3.1.2
MarkupSafe==2.1.5
openpyxl==3.1.2
psycopg2-binary==2.9.6
pydantic==1.10.18
PyJWT==2.9.0
python-dotenv==1.0.0
python-multipart==0.0.5
PyYAML==6.0.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.47
starlette==0.27.0
typing-extensions==4.12.2
uvicorn==0.22.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
